problem 8 implemented. started on problem 9

+++ p001_020.h
@@ -4,17 +4,17 @@
 #include <math.h>
 #include "common_utls.h"
 
-int p001();
-int p002();
-int p003();
-int p004();
-int p005();
-int p006();
-int p007();
-int p008();
+long p001();
+long p002();
+long p003();
+long p004();
+long p005();
+long p006();
+long p007();
+long p008();
 
 long p001_020_expected(int index);
 
-int (*p001_020_select(int index))();
+long (*p001_020_select(int index))();
 
 #endif
